Basic implementation of self-pipe trick
Ideally the network thread's select() call will block until there are new incoming data. However, this makes it impossible to send data (from the network thread anyway).

We want to send data from the network thread because it makes all changes to our set of clients/sockets synchronous (not yet implemented) and that simplifies things.

A way to achieve what we want is using "the self-pipe trick". The idea is that the network thread listen, in its select() call, listens for incoming data on a special "wake-me-up" posix pipe. The main thread can then send data to this pipe to wake it up.

This achieves what we want:

* Few system calls (select() has no timeout)
* Commands (send, disconnect, etc.) from the main thread are processed very fast by network thread