Fixed potential race condition in ChunkRingBuffer
The ChunkRingBuffer did not copy memory, it only handed back a pointer that the user could use for whatever. After this chunk buffer read, this memory can in theory be overwritten by the writing part of the chunk buffer - and this would in effect corrupt the memory retrived by the reading user of chunk buffer. Not good. This change addresses this issue by copying the actual chunk to the user before incrementing ReadIndex.