Proper serialization of commands and events
We use to do a lot of "unsafe" pointer casting between buffers and structs. This change improves this by instead explicitly reading simple integral data types into structs.

As an added side bonus this maker it easier to compress commands/events during serialization for better memory usage (if we need it later).