Refactoring: Introduced buffer as a (addr, length) pair
The following pattern was used a lot in the code:

x(void *Buffer, memsize Length)

So much in fact, that I was getting tired of it. Partly because it was longer than necessary but also because it often introduced ambiguity. And partly, because it does not make logically sense that the "address" of a buffer actually *is* the buffer. No, it would be better if the buffer address was only one of the properties describing a buffer.

So I introduced the concept of "buffer" as strictly meaning a pair (address, length). Very explicit and hard to misunderstand. Also, it makes the code much shorter in a lot of places.

This is how I think of the code now:

Buffer address: The actual adresse of a buffer. A pointer.
Buffer length: The length of the buffer, typically a memsize.
Buffer block: A chunk stack allocated data, for example: char BufferBlock[32].
Buffer: A pair consisting of one address and one length. Formalized in the struct called buffer (see code/lib/def.h).