Simplified validation of incoming message length
I added a new concept: the packet. A packet consists of a packet header and a message. The packet header is only a 16bit unsigned integer expressing the size of the following message. In that way, the receiving end knows when they have enough data to actually unserialize and validate any received message.

This let us skip a lot of cumbersome validation code.