Order serialization uses new data writer which is backed by linear allocator
Besides being easier to use this enables us to pack memory closer together. With the old method you'd have to supply the serializers with a buffer large enough to hold it - ~95% of the buffer space would not be used because most orders aren't large. That leaves whole in "hot" active working memory if other serializations is to be done at the same time. With this new method other serializations would start off at the address the first old stopped thus packing memory closer together (higher cache hit rate).